Reimplemented merge functions in service
