+++ OrderApi/Repository/IOrderRepository.cs
@@ -7,7 +7,4 @@ public interface IOrderRepository : IManageable<Order, Order>
 {
     Task<PaginatedResult<Order>> GetAllPaginatedAsync(int pageNumber, int pageSize, string searchTerm, Filter? filter,
         Sort? sort);
-
-    Task<CollectionSnippet<OrderDetailsSnippet>> GetAllByUserIdAsync(Guid id);
-    Task<SingleSnippet<OrderCardSnippet>> GetCardSnippetByUserIdAsync(Guid id);
 }
