+++ OrderApi/Services/IOrderService.cs
@@ -1,4 +1,5 @@
 ﻿using Library.Common;
+using Library.DTOs.UserRelated.User;
 using Library.Interfaces;
 
 namespace OrderApi.Services
@@ -7,5 +8,7 @@ namespace OrderApi.Services
     {
         Task<PaginatedResult<OrderDto>> GetAllAsync(int pageNumber, int pageSize, string searchTerm, Filter? filter, Sort? sort);
         Task DeleteAsync(Guid id);
+        Task<SingleSnippet<OrderCardSnippet>> GetCardSnippetByUserIdAsync(Guid id);
+        Task<CollectionSnippet<OrderDetailsSnippet>> GetAllByUserIdAsync(Guid id);
     }
 }
