+++ OrderApi/Services/OrderService.cs
@@ -1,16 +1,24 @@
 ﻿using AutoMapper;
+using BookAPI;
+using BookAPI.Repositories;
+using BookAPI.Repositories.Interfaces;
 using Library.Common;
+using Library.DTOs.UserRelated.User;
+using Library.Interfaces;
+using Library.Sorts;
+using Microsoft.EntityFrameworkCore;
 using OrderApi.Models;
+using OrderAPI;
 
 namespace OrderApi.Services
 {
-    public class OrderService(IOrderRepository repository, IMapper mapper, ILogger<IOrderService> logger) : IOrderService
+    public class OrderService(IOrderRepository repository, IMapper mapper, ILogger<IOrderService> logger, IBookRepository bookRepository) : IOrderService
     {
         private readonly IOrderRepository _repository = repository;
         private readonly IMapper _mapper = mapper;
         private readonly ILogger<IOrderService> _logger = logger;
         private string _message = string.Empty;
-
+        private readonly IBookRepository _bookRepository = bookRepository;
         public async Task<PaginatedResult<OrderDto>> GetAllAsync(int pageNumber, int pageSize, string searchTerm, Filter? filter, Sort? sort)
         {
             var paginatedOrders = await _repository.GetAllPaginatedAsync(pageNumber, pageSize, searchTerm, filter, sort);
@@ -122,6 +130,73 @@ namespace OrderApi.Services
                 _message = $"Error occurred while deleting the order with ID [{id}].";
                 _logger.LogError(_message);
                 throw new InvalidOperationException(_message, ex);
+            }
+        }
+
+        public async Task<SingleSnippet<OrderCardSnippet>> GetCardSnippetByUserIdAsync(Guid id)
+        {
+            try
+            {
+
+                OrderFilter filter = new OrderFilter() { UserId = id };
+                OrderSort sort = new OrderSort() { OrderDate = Bool.DESCENDING };
+
+                PaginatedResult<Order> orders = await _repository.GetAllPaginatedAsync(GlobalConstants.DefaultPageNumber,1,"",filter,sort);
+
+
+
+                var orderSnippet = new OrderCardSnippet();
+
+                if (orders.TotalCount > 0)
+                {
+                    orderSnippet = new OrderCardSnippet
+                    {
+                        LastOrder = orders.Items.First().OrderId.ToString().Split('-')[4],
+                        OrdersCount = orders.TotalCount
+                    };
+                }
+
+                return new SingleSnippet<OrderCardSnippet>(false, orderSnippet);
+            }
+            catch
+            {
+                return new SingleSnippet<OrderCardSnippet>(true, new OrderCardSnippet());
+            }
+        }
+
+        public async Task<CollectionSnippet<OrderDetailsSnippet>> GetAllByUserIdAsync(Guid id)
+        {
+            try
+            {
+                var filter = new OrderFilter { UserId = id };
+
+                var orders = await _repository.GetAllPaginatedAsync(GlobalConstants.DefaultPageNumber, GlobalConstants.DefaultPageSize, "", filter,null);
+                var orderDetailsSnippets = new List<OrderDetailsSnippet>();
+
+                foreach (var order in orders.Items)
+                {
+                    var bookNames = new List<string>();
+
+                    foreach (var book in order.Books)
+                    {
+                        var bookObject = await _bookRepository.GetByIdAsync(book.Key);
+                        if (bookObject != null)
+                            bookNames.Add(bookObject.Title);
+                    }
+
+                    orderDetailsSnippets.Add(new OrderDetailsSnippet
+                    {
+                        OrderUiId = order.OrderId.ToString().Split('-')[4],
+                        Price = order.Price + order.DeliveryPrice,
+                        BookNames = bookNames
+                    });
+                }
+
+                return new CollectionSnippet<OrderDetailsSnippet>(false, orderDetailsSnippets);
+            }
+            catch
+            {
+                return new CollectionSnippet<OrderDetailsSnippet>(true, new List<OrderDetailsSnippet>());
             }
         }
     }
